+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- `Logtail` - sends json pretty formatted
+- `Logtail`
+  - sends json pretty formatted
+  - setUrl() instead of setUri() (old method is deprecated)
 
 ## [1.2.5](https://github.com/orisai/nette-monolog/compare/1.2.4...1.2.5) - 2022-12-14